+++ src/blocktype.h
@@ -34,7 +34,7 @@ Boston, MA 02111-1307, USA.
   int elsize;                                                        \
   type *tempel
 
-void *Blocktype_newf (int elsize);
+void *Blocktype_newf (size_t elsize);
 void Blocktype_allocf (void *b);
 void Blocktype_free (void *bbb, void *el);
